change return type of getDistance + implement euclidean dist of matching points

+++ quadcopter/main.cpp
@@ -9,7 +9,7 @@
 #include <chrono>
 #include <thread>
 
-int calculateDistance(cv::Mat newImg, cv::Mat oldImg);
+double calculateDistance(cv::Mat newImg, cv::Mat oldImg);
 
 
 int main(int argc, char **argv) {
